fix(kesagent): compare the declared frame length without a narrowing cast
readFrame compared a uint32 length prefix against an int maxSize by
narrowing maxSize to uint32 under a gosec suppression. Widening both sides
to int64 converts losslessly and makes the bound exact for any maxSize, so
the suppression is no longer needed.

+++ kesagent/protocol.go
@@ -170,9 +170,9 @@ func readFrame(r io.Reader, maxSize int, v any) error {
 	if n == 0 {
 		return errors.New("kesagent: zero-length frame")
 	}
-	if n > uint32(
-		maxSize,
-	) { // #nosec G115 -- maxSize is always a small positive constant
+	// Widened to int64 rather than narrowing maxSize to uint32: both sides
+	// convert losslessly, so the bound is exact for any maxSize.
+	if int64(n) > int64(maxSize) {
 		return fmt.Errorf(
 			"%w: declared %d bytes > %d",
 			errFrameTooLarge,
